Use requestAnimationFrame to position modal opening (#514)
This helps to mitigate some of the glitchy scrolling where the modal overlay opening would visible jump around on scroll. It does not totally fix the problem, but is much better.

A possible way to improve things further would be to position with CSS transforms, rather than changing x and y.