Refactor styleVariables, remove automatic calculations
This removes polished as a dependency, changes the variable names, and requires the user to explicitly define all styles for each state. We were calculating sane defaults, based on other variables you passed, but that was buggy, so setting them all is best.

This will save us ~4 kb from our bundle size as well.