Prevent scrolling when element receives focus after render
When scrollTo is configured with an scrollIntoView object that has behavior set to
'smooth', the element's focus can hijack the scroll and interrupt the target from
scrolling fully into view.

This is particularly noticeable in browsers such as Chrome/Chromium, that implement a
slower transition for the smooth scroll or when the distance to scroll is far.

Configuring the element focus with the preventScroll option will ensure the smooth
scroll transition will not be interrupted.